requirements.txt
pandas>=1.5.1
python-dateutil>=2.8.2
python-dotenv==1.0.1
sqlalchemy>=1.4.0